Updated import paths for moved components
